requirements.txt
bleach==1.4.1
django-crispy-forms==1.5.0
django-ipware==1.1.0
django-reversion==1.9.3
Django==1.8.3
djangocodemirror==0.9.7
git+https://github.com/django-ckeditor/django-ckeditor@cefb310ccfafdcee05b1f2be2b5bb9ccd215b107#egg=django-ckeditor
gunicorn==19.3.0
ipython==4.0.0
loremipsum==1.0.5
ndg-httpsclient==0.4.0
Pillow==2.9.0
pyasn1==0.1.8
pyjade==3.1.0
pyOpenSSL==0.15.1
requests-oauthlib==0.5.0